New 'common' sub pkg with ifaces.go & datapoint.go
ifaces.go (new file):
A bunch of interfaces that will be common. Had to be placed here
because the type system in Go is a bit wonky, description for why
is placed at the top of the file. Interfaces:
- VecContainer
- PayloadContainer
- PayloadReceiver
- PayloadDrainer
- PayloadExpirer
- MemTrimmer
- VecMover
- PayloadDistributer
- KNNSearcher
- DataContainer

datapoint.go (new file):
For storing the smallest type of meaningful data in the project;
DataPoint (with vec, expiration and payload). It implements the
'PayloadContainer' interface listed above (ifaces.go section).