Del DataPoint type and substitute with DataPoint iface ++
Decided that playing these naming convention games is
unproductive here, i'll call a spade a spade. So
completely deleted the common/datapoint.go file
and changed the PayloadContainer interface in
common/ifaces.go to DataPoint. This required a few
naming changes in other interfaces + docs had to
be tweaked as well.