Centroid implementation.
Centroid type that implements the interface with the same name
in pkg/kmeans/common, so a bunch of methods in and tests in
the following files:

centroid.go (new file):
Concrete Centroid type, with a factory methods and dedicated
arg type. Methods are
- (*Centroid) AddDataPoint
- (*Centroid) rmDataPoint
- (*Centroid) dataPointVecGenerator
- (*Centroid) DrainUnordered
- (*Centroid) DrainOrdered
- (*Centroid) Expire
- (*Centroid) LenDP
- (*Centroid) MemTrim
- (*Centroid) MoveVector
- (*Centroid) DistributeDataPoints
- (*Centroid) KNNLookup

.. all have fairly good, and at time lengthy,
 documentation -- that isn't included in
this commit msg.

centroid_test.go (new file):
Tests for almost all the methods listed above.
Some aren't covered, like LenDP, as that is
hard to mess up.. probably.