Add auto-adjust Centroid.vec when adding/removing dps.
Realised that calling Centroid.MoveVector periodically,
or god-forbid every time one adds/removes a datapoint,
would be super-inefficient... obviously...

centroid.go (changed a few methods):
..So changed Centroid.rmDataPoint to do this and hope
to remember to allways use this method when actually
removing datapoints (checked the file, looks like that
is the case). Also added Centroid.addDataPoint and
hooked it into the exported method with the same name.

...Also changed Centroid.MemTrim for a similar reason.
Plus, i realised that the implementation of this method
made the use-case utterly pointless: It is supposed to
reduce the internal DataPoint cap, but it created a
new one with the same cap...

centroid_test.go (add tests):
Added tests for Centroid.addDataPoint and
Centroid.rmDataPoint.