complete 58fc2c3 (that commit was unfinished)
Did a blunder, wrote in 58fc2c3 that 19d71e was applied
to kmeans/centroidmanage (auto-adjusting internal
vector, as opposed to having to call
CentroidManager.MoveVector() each time some change
was applied to centroids/datapoints), but that was
not completely the case -- that 58fc2c3 commit only
auto-adjusted the CentroidManager for AddDataPoint.

This commit actually applies automatic vector adjustment
for pretty much all methods that change state, which
is a major change, so won't list method names here.

Also tweaked test for the changes.