bugfix + new test.
bug + resolution:
There was a hidden bug in CentroidManager.KNNLoopkup; if drain
was on and the result requirement was filled (len of res), then
it would simply return from the inner loop that iterated over
datapoints. The issue here is that if drain is on, then the
drained centroids would give some dps to iterate over, and
if an early return is executed, then those dps would be lost
to the GC. The fix was to put those dps back into the drained
centroid instead of simply returning.

new test:
Realised the method mentioned above (KNNLookup) was not tested
in cmanager_test.go, so added a func for that.