stopped interface madness
Hit a roadblock while doing the kmeans/rpc pkg, the common.DataPoint
interface was difficult to work with; using that with rpc in Go
doesn't seem to be possible and converting it into a concrete
datapoint thing wasn't feasible without some changes to that
dp interface (needed hidden->public fields). So decided to just
drop the interface overuse in the kmeans pkg, it was counterproductive
and pointless for the most part (made comments a lot more vague
as well). This naturally required a lot of changes in these files
 - modified:   centroid/centroid.go
 - modified:   centroid/centroid_test.go
 - modified:   centroidmanager/cmanager.go
 - modified:   centroidmanager/cmanager_test.go
 - new file:   common/datapoint.go
 - deleted:    common/ifaces.go