add core/nodes/ (find 'best' addrs, given a vec)
These are some tools that finds remote nodes (pkg/kmeans/rpc)
that are most similar to a vec. Intent/purpose is that it seems
highly likely at this point that it would be useful for an event
loop thing (not yet impl) and/or the main public-facing api for
this project (also not yet impl).