change DistributeDataPoints dramatically
There were a couple of issues with the previous usage
of DistributeDataPoints (client and server);
  1: The heavy load was done on the client side, kinda,
     it fetched dps, then sent them away.. which is bad.
  2: Used only CentroidManager.Vec to find best-fits, so
     not necessarily accurate.

The fix to problem 1 was to put the code on the server side,
so node A tells node B to drain dps and then distribute them
away to some other places, without necessarily sending anything
to node A.

Fix to problem 2 was to have two variants (DistributeDataPointsFast
and DistributeDataPointsAccurate), one using kmeansClient.Vec()
(so CentroidManager granularity) and the other using
kmeansClient.NearestCentroidVec (very granular but slow).

serverutils.go was added so server.go didn't get too messy,
it currently has funcs for fetching vecs, as desribed above.

also changed rpc_test.go for the new stuff.