improved tMonitor.pollAccuracy and some other stuff
Previously, the tMonitor.pollAccuracy method used
dps that were global to the file that were pre-loaded...
for some reason. This is no longer the case, as
query vectors can just be pulled randomly from the
network.

Also changed dpAddRand so the bias can be toggled with
configuration variables that are declared at the top
of the file.