add build flag in el_test.go + a bit cleanup
Added buildflag "monitor" so the monitor won't run automatically
and implicitly when 'go test' is ran on the entire project, since
it's not really a test like that.

Also finally bothered creating g_addrs (slice of addrs for the
monitor session) using a loop... should probably have done that
earlier. As such, had to move the 'g_network' (test network)
instantiation (is that a word in Go?) into the new init func
used for the aforementioned address creation.