rm rpc addrs from JSON opt, moved to cfg and handler T.
A bit of a bother having to specify all addrs for the rpc
network in a user-facing NAPI, so moved that from the
JSON options -> the APIConfig and handler types. This
made it necessary to remove some stuff from the test.

Also redundant code (specifically []string <-> []Addr)
conversion/helper funcs.